requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
html5lib==1.1 